+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- You can now use custom identifiers for entity objects.
+  You can access these from World instances with `entity = world[uid]`.
+
+### Removed
+- Dropped support for unpickling v1.0 World objects.
 
 ## [1.2.0] - 2023-04-26
 ### Added